Move useServiceConstructor to core package
