+++ libs/@blockprotocol/hook/src/react.ts
@@ -1,80 +1,7 @@
-import { ServiceHandler } from "@blockprotocol/core";
-import { GenericMessageCallback } from "@blockprotocol/core/src/types";
-import { RefObject, useEffect, useLayoutEffect, useRef, useState } from "react";
+import { useServiceConstructor } from "@blockprotocol/core/dist/esm/react.js";
+import { RefObject, useLayoutEffect, useRef, useState } from "react";
 
 import { HookBlockHandler, HookEmbedderHandler } from "./index.js";
-
-type ServiceConstructor<T extends ServiceHandler> = {
-  new (arg: {
-    callbacks?: Record<string, GenericMessageCallback>;
-    element?: HTMLElement | null;
-  }): T;
-};
-
-const useServiceConstructor = <T extends ServiceHandler>({
-  Handler,
-  constructorArgs,
-  ref,
-}: {
-  Handler: ServiceConstructor<T>;
-  constructorArgs?: { callbacks?: Record<string, GenericMessageCallback> };
-  ref: RefObject<HTMLElement>;
-}) => {
-  const previousRef = useRef<HTMLElement | null>(null);
-  const initialisedRef = useRef(false);
-
-  const [service, setService] = useState<T>(
-    () => new Handler(constructorArgs ?? {}),
-  );
-
-  const previousCallbacks = useRef<Record<
-    string,
-    GenericMessageCallback
-  > | null>(null);
-
-  useLayoutEffect(() => {
-    if (previousCallbacks.current) {
-      service.removeCallbacks(previousCallbacks.current);
-    }
-
-    previousCallbacks.current = constructorArgs?.callbacks ?? null;
-
-    if (constructorArgs?.callbacks) {
-      service.registerCallbacks(constructorArgs.callbacks);
-    }
-  });
-
-  // eslint-disable-next-line react-hooks/exhaustive-deps -- will not loop & we don't want to reconstruct on other args
-  useEffect(() => {
-    if (ref.current === previousRef.current) {
-      return;
-    }
-
-    if (previousRef.current) {
-      service.destroy();
-    }
-
-    previousRef.current = ref.current;
-
-    if (ref.current) {
-      if (!initialisedRef.current) {
-        service.initialize(ref.current);
-      } else {
-        setService(
-          new Handler({
-            element: ref.current,
-            ...constructorArgs,
-          }),
-        );
-      }
-
-      initialisedRef.current = true;
-    }
-  });
-
-  return service;
-};
-
 /**
  * Create a HookBlockHandler instance, using a reference to an element in the
  * block.
